+++ EOLib/PacketHandlers/Spell/SpellRequestHandler.cs
@@ -5,18 +5,21 @@ using EOLib.Net;
 using EOLib.Net.Handlers;
 using System.Collections.Generic;
 
-namespace EOLib.PacketHandlers
+namespace EOLib.PacketHandlers.Spell
 {
+    /// <summary>
+    /// Sent when a player starts a spell chant
+    /// </summary>
     [AutoMappedType]
-    public class PlayerStartSpellCastHandler : InGameOnlyPacketHandler
+    public class SpellRequestHandler : InGameOnlyPacketHandler
     {
         private readonly IEnumerable<IOtherCharacterAnimationNotifier> animationNotifiers;
 
         public override PacketFamily Family => PacketFamily.Spell;
         public override PacketAction Action => PacketAction.Request;
 
-        public PlayerStartSpellCastHandler(IPlayerInfoProvider playerInfoProvider,
-                                           IEnumerable<IOtherCharacterAnimationNotifier> animationNotifiers)
+        public SpellRequestHandler(IPlayerInfoProvider playerInfoProvider,
+                                   IEnumerable<IOtherCharacterAnimationNotifier> animationNotifiers)
             : base(playerInfoProvider)
         {
             this.animationNotifiers = animationNotifiers;
